More Sink configuration options
