+++ src/Serilog.Sinks.YandexCloud/Serilog.Sinks.YandexCloud.csproj
@@ -4,8 +4,8 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <AssemblyVersion>2.0.0</AssemblyVersion>
-        <Version>2.0.0</Version>
+        <AssemblyVersion>2.0.2</AssemblyVersion>
+        <Version>2.0.2</Version>
         <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
         <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
         <RepositoryType>git</RepositoryType>
